took out create playlist calls to api module, currently broken with spotify login
