+++ routes/index.js
@@ -1,12 +1,10 @@
 var express = require('express');
 var router = express.Router();
 var querystring = require('querystring');
-var spotify = require('../lib/spotify.js');
-var generator = require('../lib/generator');
+var spotify = require('../lib/spotify');
+var api = require('../lib/api');
 var url = require('url');
-var extend = require('util')._extend;
-var bluemix = require('../lib/bluemix.js');
-var watson = require('watson-developer-cloud');
+
 var fs = require('fs');
 var multer = require('multer');
 var storage = multer.diskStorage({
@@ -18,16 +16,6 @@ var storage = multer.diskStorage({
   }
 })
 var upload = multer({ storage: storage });
-
-// if bluemix credentials exists, then override local
-var credentials = extend({
-  username: '0e8ac13c-e154-4cf8-be34-4f88ca9cac47',
-  password: 'tFYIUlzIFylf',
-  version: 'v1'
-}, bluemix.getServiceCreds('visual_insights')); // VCAP_SERVICES
-
-// wrapper for watson visual insights
-var visual_insights = watson.visual_insights(credentials);
 
 // Either login or query
 router.get('/', function(req, res, next) {
@@ -92,23 +80,15 @@ router.post('/generateAndRedirect', upload.single('images_file'), function(req, 
 
       if(words){
         var words = words.split(' ');
-        generator.getPlaylistFromWords(words, access_token, user_id, function(response) {
-            res.redirect('http://open.spotify.com/user/' + user_id + '/playlist/' + response);
+        api.getPlaylistFromWords(words, access_token, user_id, function(playlistID) {
+            res.redirect('http://open.spotify.com/user/' + user_id + '/playlist/' + playlistID);
         });
       }
       if(file){
         var images_file = fs.createReadStream(file.path);
         if(images_file){
-          visual_insights.summary({images_file: images_file}, function (err, result) {
-            if(err){
-              next(err);
-            }
-            else{
-              var descriptors = result.summary;
-              generator.getPlaylistFromDescriptors(descriptors, access_token, user_id, function(response) {
-                res.redirect('http://open.spotify.com/user/' + user_id + '/playlist/' + response);
-              });
-            }
+          api.getPlaylistFromImages(images_file, access_token, user_id, function(playlistID) {
+              res.redirect('http://open.spotify.com/user/' + user_id + '/playlist/' + playlistID);
           });
         }
       }
